huge cleanup infra code
