+++ gatling-infra/src/main/java/nl/codecontrol/cdk/gatling/monitoring/FargateTaskRole.java
